Add files via upload
In this notebook, I developed an algorithm that could be used as part of a mobile or web app. The code will accept any user-supplied image as input. If a dog is detected in the image, it will provide an estimate of the dog's breed. If a human is detected, it will provide an estimate of the dog breed that is most resembling